requirements.txt
flask==3.0.3
ultralytics==8.3.179
pillow==10.4.0


